[SYCL][ESIMD] Add support for generalized raw send (#15291)
This change adds support for the `raw_sendg` and `raw_sendgc`
instruction which will be supported on Xe3+ hardware. The documentation
for the instruction is
[here](https://literate-invention-48704e9f.pages.github.io/instructions/RAW_SENDG.html).
The change adding support in VC is
[here](https://github.com/intel-innersource/drivers.gpu.unified/commit/64d0a0fb83e44461b810b2e30b1b06268d8a04e4).

There is no VC intrinsic for this instruction, so we were recommended to
use inline assembly by the VC team. I will ask Victor to review the
inline assembly as well. I manually tested this on a manually-downloaded
very recent IGC driver, but we don't have that driver in CI here nor do
we have the hardware to run an E2E test, so this is only tested with a
compile-time test for now.

I was not able to think of a way to reduce the `if constexpr` conditions
based on the `isSendc` and `isEOT` arguments in the inline assembly
implementation, let me know if you have an idea.

We have 12 overall variants, all named `raw_sendg` and overloaded.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>