[NFC] Use options stored in CodeGenTBAA instead of CodeGenModule (#16054)
Intel has added some code for TBAA to handle bitfields which requires
a CodeGenModule object. In order to get that for TBAA it is added to
the CodeGenTBAA object. Unfortunately for non-standard compiles (such
as clang-repl), that CodeGenModule object is not fully formed and the
LangOpts and CodeGenOpts are not valid and cannot be used.

Fortunately these options are already correctly set up in the
CodeGenTBAA so we can use them from there instead of from the CGM.

The other use of the GGM in TBAA is protected by Intel customization
options so it should be safe from non-standard uses.

Fixes: CMPLRLLVM-51298