Extend Intel_PartialInliner pass to recognize function form when PGO is used (#16075)
Some performance is missed when PGO (instrumented or sampled) is used on
511.povray because a transformation done for 'base' was not occuring
when
PGO data was used due to PGO changing the IR prior to the
Intel_PartialInliner
pass run. Specfically:
1. PGO has identified indirect function call targets, and put in
specialized
calls to them, elarging the number of basic blocks, so the limit is
increased.
2. PGO has marked hot functions with a hint to the inliner that they
would be
given priority for inlining. Removed this guard from this pass, because
the
   the intent of this pass is to make the function inlinable anyway.

[CMPLRLLVM-49898]