[VecClone][VPlanCallVecDecisions] Scalarize vector function calls for unoptimized uval args (#15597) (#16039)
Linear uval args are not currently supported at -O0 for two reasons:
1) The unoptimized IR presents problems for VecClone in that it must
somehow
determine how the uval ptrs are tracked through memory and make the
appropriate
   "fake" stores using that information.
2) The way in which VecClone applies stride is incorrect for unoptimized
IR and
   this should be fixed as discussed in CMPLRLLVM-49320.

For now, this PR teaches VecClone and CallVecDecisions to just scalarize
vector functions
that have uval args and are compiled using the optnone attribute. The
optnone attribute tells
us that incoming IR is not optimized (e.g., we have loads/stores of args
through memory).