[OpenMP] Introduce dynamic reduction buffers (#15847)
Atomic-free teams reduction buffer sizes by design
depend on the actual number of teams a target region
is run with.

The original approach was to rely on the
static approach, i.e. having a compiler option to
provide the number of teams hoping the actual one wouldn't exceed
the value of that option.

This change introduces a proper dynamic approach, when the runtime
adjusts the size of a mapitem marked with a special maptype
by the actual number of teams.