[OMP][Offload] Support -ftarget-compile-fast for JIT (#16112)
This option is already supported for OMP Offload AOT and SYCL JIT and
AOT.

Implement the missing piece similar to what was done for SYCL in
https://github.com/intel/llvm/commit/92855bea87f452ae4b16ed4872c16669ce80ba86#

Basically we just check if it's Intel GPU and if so convert
-ftarget-compile-fast passed from the driver through image opts to the
real backend option. If it's not an Intel GPU, just drop it from options
as to not cause an error on other hardware.

I needed to get the vendor ID to check if it's an Intel GPU and the OCL
plugin didn't seem to do that yet, so I added it. L0 already had it.

Jira: CMPLRLLVM-47050

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>