[VPlan][NFCI] Implement OptReportTraits<vpo::VPLoop>::getDebugLoc() (#15965)
Implement the getDebugLoc() interface for VPLoop to properly capture
loop start information for the opt report.  Modify existing VPLoopInfo
interfaces to capture the data when creating or copying loop information
so that any future uses will not miss the requirement.  Capture the
LoopID metadata when building the CFG for the HIR path.  Remove previous
tactical solution to attach debug location to the op report during code
generation.

Reference: CMPMLRLLVM-50802