[GVBasedMultiversioning] Support scoped multiversioning. (#16059)
This allows GVMV to be applied only to a portion of function. It's used
to
deal with instructions that may modify the global, but can't be excluded
with the approach of adding a new invariant. It's particularly useful
against unknown side-effects at the entry or exit of a function.

To keep the patches short, this patch only adds the definition of scope
and
some basic analysis to identify a scope for multiversioning. Scope
support
for the cloning transformation will come later.

JIRA: CMPLRLLVM-49723