[NFC][VecClone] Fix bug with handling of elementsize intrinsic (#16227)
Previous behavior of storing the corresponding element size of a
pointer argument was flawed in that it used a reverse VMap of the
original arg to the new arg created during cloning the function.
During linear arg processing, the new arg value is replaced with
a store of the original arg to memory and then loaded back. This
loaded value is then used as the new arg replacement. Due to this
RAUW, this memory may no longer be available and led to some flaky
behavior. The new approach is to set aside a slot in a vector for
each argument to store the element size. If the arg is not a pointer
this slot is nullptr. This way we can easily get the arg number from
the function signature and look up the corresponding element size.