Resolve of merge (conflict) cbdccb30c23f [RA] Split a virtual register in cold blocks if it is not .. 'main' -> 'xmain-web' (#16319)
CONFLICT (content): Merge conflict in
llvm/test/CodeGen/X86/speculative-load-hardening.ll

  commit cbdccb30c23f71f20d05b19256232419e7c5e517
  Author: Guozhi Wei <carrot@google.com>
  Date:   Fri Sep 15 19:52:50 2023 +0000

[RA] Split a virtual register in cold blocks if it is not assigned
preferred physical register

If a virtual register is not assigned preferred physical register, it
means some
COPY instructions will be changed to real register move instructions. In
this
case we can try to split the virtual register in colder blocks, if
success, the
original COPY instructions can be deleted, and the new COPY instructions
in
colder blocks will be generated as register move instructions. It
results in
      fewer dynamic register move instructions executed.

The new test case split-reg-with-hint.ll gives an example, the hot path
contains
24 instructions without this patch, now it is only 4 instructions with
this
      patch.

      Differential Revision: https://reviews.llvm.org/D156491

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Tyler Lanphear <tyler.lanphear@intel.com>