[VPlan][VSLP] Switch all utilities to work with VPValue vectors. (#16258)
With the change every VSLP utility that works with vectors now takes
vector
of VPValues on input as contrary to VPInstructions. W/o the change we
have a
mix: there are utilities that take VPValues and there are utilities that
that VPInstructions. This is a step towards encapsulating VPValues
vector
into SLP Node new data type and let the utilities to take the new data
type
on input or become members of the new data type.

There is an improvement in detecting base instruction node while
searching
for vectorizable subset of VPInstructions within vector of VPValues
(in getVecInstsVector()). However it is yet NFC change for now until VL
reduction is not supported and a vector of VPInstructions is discarded
regardless if it is shorter than input VPValue vector.

The change allowed to merge two estimateSLPCostDifference() methods that
we had
before different implementations for VPValues and VPInstructions on
input.