This is to merge communit https://reviews.llvm.org/D157452 (#16304)
Author: Alex Voicu <alexandru.voicu@amd.com>
Date:   Mon Aug 28 20:44:06 2023 +0100

[Clang][CodeGen] `typeid` needs special care when `type_info` is not in
the default AS

After https://reviews.llvm.org/D153092, for targets that use a
non-default AS for globals, an "interesting" situation arises around
typeid and its paired type, type_info:

- on the AST level, the type_info interface is defined with default /
generic addresses, be it for function arguments, or for this;
- in IR, type_info values are globals, and thus pointers to type_info
values are pointers to global

This leads to a mismatch between the function signature / formal type of
the argument, and its actual type. Currently we try to handle such
mismatches via `bitcast`, but that is wrong in this case, since an
`ascast` is required. This patch ensures that iff the pointer to
`type_info` points to a non-default AS, an ascast is inserted so as to
match the `typeid` interface / return value type.

    Reviewed by: yaxunl

    Differential Revision: https://reviews.llvm.org/D157452