Resolve of merge (conflict) b1e3cd1d7944 [driver] Conditionally include installed libc++ headers fo.. 'main' -> 'xmain-web' (#16344)
CONFLICT (content): Merge conflict in
clang/lib/Driver/ToolChains/Gnu.cpp

  commit b1e3cd1d79443603dc003441e07cdd8d30bb7f26
  Author: Shoaib Meenai <smeenai@fb.com>
  Date:   Wed Aug 30 15:11:07 2023 -0700

[driver] Conditionally include installed libc++ headers for Android

https://reviews.llvm.org/D71154 prevented Clang from search for libc++
      headers installed alongside the driver when targeting Android. The
motivation was the NDK's use of a different libc++ inline namespace
(`__ndk1` instead of the standard `__1`), which made regular libc++
      headers incompatible with the NDK's libc++ library.

Since then, libc++ has gained the ability to install its `__config_site`
header (which controls the inline namespace, among other things) to a
per-target include directory, which enables per-target customizations.
If this directory is present, the user has expressly built libc++ for
      Android, and we should use those headers.

The motivation is that, with the current setup, if a user builds their
own libc++ for Android, they'll use the library they built themselves
but the NDK's headers instead of their own, which is surprising at best
and can cause all sorts of problems (e.g. if you built your own libc++
with a different ABI configuration). It's important to match the headers
and libraries in that scenario, and checking for an Android per-target
include directory lets us do so without regressing the original scenario
      which https://reviews.llvm.org/D71154 was addressing.

While I'm here, switch to using sys::path::append instead of slashes
directly, to get system path separators on Windows, which is consistent
with how library paths are constructed (and that consistency will be
important in a follow-up, where we use a common search function for the
      include and library path construction).

(As an aside, one of the motivations for https://reviews.llvm.org/D71154
      was to support targeting both Android and Apple platforms, which
expected libc++ headers to be provided by the toolcain at the time.
Apple has since switched to including libc++ headers in the platform SDK
      instead of in the toolchain, so that specific motivation no longer
      applies either.)

      Reviewed By: phosek

      Differential Revision: https://reviews.llvm.org/D159292

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Tyler Lanphear <tyler.lanphear@intel.com>