[NFC][VPlan][SLP] Remove NOP sort. (#16334)
The code which does sorting of memory references w/o disturbing lexical
order
is NOP sort, i.w. it doesn't change input order.
We need to take a risk and do sort regardless of the lexical order or
add
checks for dependencies. Either way the current code does not change the
order
of memrefs in iput array and can be disabled or deleted.