[HIRSSADeconstruction] More robust checks for reordering to keep IV livein copy at the end (#16286)
There is a logic to keep the IV livein copy at the end of the loop when
adding other livein copies which may use IV as a blob to prevent live
range issues.

It was being triggered for cases where the livein copy doesn't really
represent the IV. Its AddRec form may belong to an inner loop or it
may be non-linear function of the IV. This PR makes he checks more
robust. We also skip the reordering checks if the current copy is known
to belong to AddRec phi. 

CMPLRLLVM-51560