CMPLRLLVM-51642: Check for no linear ref when the ref is in the max loop level during HIROptVarPredicate (#16403)
The analysis process in HIROptVarPredicate produces an assertion
if the ref is inside the max innermost loop level (loop nest 9).
We are trying to analyze for invariance in level 10, but it doesn't
exists. This patch skip the check if invariance when we are at max loop
level, but still checks if the ref is linear.