[NFC] [DTrans] Code cleanup of AOS-to-SOA pass pointer handling (#16441)
Now that only opaque pointers are present for the AOS-to-SOS, some
checks are no longer necessary because pointers will always
be opaque. Remove the checks to reflect that.

The community progresses in removing GetPointerTo(<ty>) calls
now that opaque pointers are the only pointer type. Remove
uses in DTrans AOS-to-SOA.

[CMPLRLLVM-51862]