[OpenCL][clang] Remove opencl.compiler.options metadata support (#16167)
CMPLRLLVM-50666. The metadata became empty since switching to open
source opencl-clang in #10997. Open source version doesn't pass
'-cl-spir-compile-options' to clang because it isn't a standard OpenCL
compile option. Nevetherless, we're not aware that #10997 is cause any
regressions that are related to the metadata change.

So this PR just remove the support in clang. For remaining backend
issues in CMPLRLLVM-12895 to interpret attribute rather than compile
option, we'll fix them soon.

---------

Co-authored-by: Mike Rice <michael.p.rice@intel.com>