Fix clang-offload-wrapper performance (#16431)
In clang-offload-wrapper, we use llvm-objcopy to copy multiple images
into a single wrapped file. Currently, we invoke llvm-objcopy once for
each image. When the number of images is large, the number of
llvm-objcopy is also large and hence we end up spending a lot of time in
just invoking this tool.
However, llvm-objcopy does allow for multiple images to be passed in
during the same invocation. We can exploit this to add all images using
a single invocation of llvm-objcopy. This reduces the compile time
drastically.

Follow-up: Adding ALL images at once results in 'too many arguments'
error. So, limiting number of images written using one llvm-objcopy copy
to 1000 (seems reasonable).

Thanks
Signed-off-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>

---------

Signed-off-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>