[SPGO] Add lto-sample-profile option for lld-link (#16464)
The pass pipeline for the LTO phase has some checks
on whether the configuration has the PGO action to
SampleUse when setting up the passes. On Linux, there
is an option to the linker, -lto-sample-profile, which
causes the configuration to be set. On lld-link, there
was no corresponding option, which prevented the
pipline from including these passes when using sample
profiling on Windows.

This PR adds the option, and sets up the LTO
configuration with it on Windows to improve performance
of SPGO with LTO on Windows.

[CMPLRLLVM-49898]