[SYCLTransforms] Fix out-of-range arg index in PipeSupport (#16498)
CMPLRLLVM-51932
Unused argument in non-kernel function may be eliminated, e.g. by
argpromotion pass. Then it isn't easy to know if an argument is pipe.
In this case it is ok to over-estimate the number of pipes. The number
is used to initialize the pipe array size. Actually used pipe is stored
to the array during runtime. Over-estimated size doesn't hurt stability.

Remove outdated comment from pass pipeline.