[CoerceTypes][SGSizeCollectorIndirect][VFAnalysis] Fix for CMPLRLLVM-35340 (#16368)
This patch fixes an issue where we reach an assert due to attempted
widening of byval pointer args. This is not allowed by the LLVM
verifier and vectorization of byval args is not yet supported. This
patch prevents any vector attributes from being updated to the global
variable for indirectly called functions and for calls to __intel_
indirect_call_xxx functions, thus preventing vectorization of them.
As a result, any kernels making calls to __intel_indirect_call_xxx
functions are also not vectorized.