CMPLRLLVM-51565: Update the bi-direction analysis when creating the DD graph (#16390)
The DD analysis considers that the bi-direction for a direction vector
needs to be created if there is any * (DVKind::ALL) before any <
(DVKind::LT). The issue is that a >= (DVKind::GE) and <= (DVKind::LE)
can represent a dependence when there is an equal value, and a greater|less
than. Therefore we need to also consider these cases for bidirection. Also,
this patch updates the function for splitting the forward and backward direction
vectors to also split when >= or <= is found.