[Loopopt] Disable recomputation for loop distribution when def chunk contains unsafe call (#16554)
CMPLRLLVM-51289

Bug occurred when we distributed a chunk that was trying to recompute a
temp which was loading a ref copied from a fake ref. Although the temp
seems linear, it's escaping via the unsafe call which was failing a
benchmark. Added a bailout for recomputation when seeing an unsafe call.