CMPLRLLVM-48084: Add support for ORing the check of multiple Ifs into one condition during HIRMemoryReductionSinking (#16485)
The HIRMemoryReductionSinking can identify a reduction inside an If
condition and sink it. The sinked instruction will be guarded inside a
condition. This patch updates the sinking process to check if multiple
reductions are inside the same condition, then the sinked reductions
will be inside one IF condition. The check for each readuction will be
compined into multiple ORs to identify if at least one happens then we
proceed to finalize the reduction.

Expected gain: polyhedron/mdbx ~1.5%