[VecClone] Fix for CMPLRLLVM-49320 (#15981)
This patch fixes a problem where stride was being applied multiple
times to uval/val arg values. The existing approach to applying
stride was flawed in that the algorithm looked for loads of the
uval/val arg value and then updated that value with the stride.
This did not account for the fact that there could be more than
one such load. The new algorithm inserts a new load of the arg
value, calculates the stride, and then stores that value as an
initial value at the beginning of the simd loop. Any subsequent
users will then be able to get the updated value with the stride
applied only once.