CMPLRLLVM-51940: Split HIRMemoryReductionSinking in two phases (#16612)
The HIRMemoryReductionSinking runs once in the pass manager. It handles
reduction inside loops. In the case of nested loops, if a reduction
inside
a condition is sunk, it will create a new condition in the parent loop.
If any of the opt-predicate passes hoist the original condition that
encloses the reduction, then new conditions that where added by  
HIRMemoryReductionSinking could have been avoided since the reductions
will become unconditional.

This change set splits the HIRMemoryReductionSinking in two phases and
the pass will be run twice in the pass manager. The first time it
handles the reductions that aren't inside an If condition. Then, the
second run will happen after all opt-predicate passes where the
conditions are clean up. Also, the second run will allow to sink any
reduction inside the conditions that are inside the loops.