[HIRCodeGen] Add logic to CodeGen invariant CEs and ref dimensions at outer loop levels (#16594)
HIRFramework propgates invariant SCEVs/closed forms into loopnests due
to the
nature of parsing. These should be cleaned up after loopopt but we don't
run
LICM pass after loopopt and adding one in cleanup passes results in
multiple
regressions. Since we have information about expr invariance in HIR, we
should
use it to generate exprs at outer loops.

This PR does that for invariant CEs and GEP ref's dimensions. This work
is
related to CMPLRLLVM-51290 but doesn't resolve the issue. We need to
extend it
for blobs. That will require some changes in SCEVExpander.