Resolve of merge (conflict) 75e648031c39 [Libomptarget] Disable AMDGPU complex math test after rece.. 'main' -> 'xmain-web' (#16693)
CONFLICT (content): Merge conflict in
openmp/libomptarget/test/offloading/std_complex_arithmetic.cpp

  commit 75e648031c39bb602b02d83021eda59fd8fc2259
  Author: Joseph Huber <jhuber6@vols.utk.edu>
  Date:   Wed Oct 4 15:23:32 2023 -0500

      [Libomptarget] Disable AMDGPU complex math test after recent patch

      Summary:
The recent patch added `-nogpulib` to make these tests only pick up what
was intentionally put into them. This had the effect of removing the
dependency on the ROCm device libs which are needed for math. This test
disables the complex math test, which is the only one that needed it,
for the time being. In the future we will implement these and provide it
      via the GPU `libm` and pass it in the same way as the GPU `libc`.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Fine, Gregory <gregory.fine@intel.com>