Resolve of merge (build) d6254e1b2e6d Introduce the initial support for OpenMP kernel language (.. 'main' -> 'xmain-web' (#16722)
commit d6254e1b2e6d1792d3a7085bd05d86adce15757e
  Author: Shilei Tian <i@tianshilei.me>
  Date:   Thu Oct 5 17:34:40 2023 -0400

      Introduce the initial support for OpenMP kernel language (#66844)

This patch starts the support for OpenMP kernel language, basically to
write
OpenMP target region in SIMT style, similar to kernel languages such as
CUDA.
What included in this first patch is the `ompx_bare` clause for `target
teams`
directive. When `ompx_bare` exists, globalization is disabled such that
local
variables will not be globalized. The runtime init/deinit function calls
will
not be emitted. That being said, almost all OpenMP executable directives
are
not supported in the region, such as parallel, task. This patch doesn't
include
the Sema checks for that, so the use of them is UB. Simple directives,
such as
atomic, can be used. We provide a set of APIs (for C, they are prefix
with
`ompx_`; for C++, they are in `ompx` namespace) to get thread id, block
id, etc.
      Please refer to
https://tianshilei.me/wp-content/uploads/llvm-hpc-2023.pdf for more
details.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Maronas, Marcos <marcos.maronas@intel.com>