[CMPLRTOOLS-21656] - Move xmain configuration logic to cmake caches. (#14166)
This adds various cmake cache files for building xmain. The primary
entrypoints are four files:
* Debug.cmake: no optimizations, all debug flags, assertions
* OptimizedDebug.cmake: optimized, -g, assertions
* Asserts.cmake: optimized, no debug flags, assertions
* Release.cmake: optimized, no debug flags, no assertions

In addition to these files, the following files can be included (in
conjunction with one of the previous files):
* Stage2.cmake: build with an existing build of the compiler
* PGOGen.cmake: set the build up for profile-guided optimization
* PGOUse.cmake: use existing profile to build the compile

These files will eventually include xmain.cmake, which provides the core
set of configuration options to configure the compiler build. It will
also, depending on the current host platform, choose to include
Linux.cmake or Windows.cmake to perform OS-specific configuration.

The file buildtype.cmake is provided to automatically dispatch to the
appropriate file(s) to include based on the value of ICS_BUILDTYPE in
the environment, intended to be used solely for automation scripts that
rely on ICS. In addition, a few other cmake caches are included to
provide support for variants that are used in automation, e.g.,
xmainxisaefi2linux.

To build a compiler with these xmain caches, it is possible to run the
following commands:

  $ cd <build dir>
  $ ics set context ws build
  $ cmake -C <workspace>/llvm/cmake/caches-intel/Debug.cmake -G Ninja \
    <workspace>/llvm/llvm
  $ ninja all deploy

The use of the "ics set context ws build" is presently necessary to capture some of the environment variables for the build (such as identifying the correct build compiler to use). Future changes will drop the requirement to run this command.